# app/microservice/public-swagger.yml
swagger: '2.0'
###############################################################################
#                              API Information
###############################################################################
info:
  title: GLobal Forest Watch API
  description: |
    The Global Forest Watch API is currently in beta. Based on feedback from trusted testers, this document specifies new API changes that are currently under development. Interested in being a trusted tester? Contact the [Data Lab](mailTo:support@wri.com) at World Resources Institute.

    All endpoints are accessible via http & https and are located at
    `https://production-api.globalforestwatch.org`.

    Forest change measures tree cover loss, tree cover gain, or forest disturbance. The following forest change data are available through the API:
    * University of Maryland tree cover loss & gain
    * FORMA alerts
    * IMAZON SAD alerts
    * Terrai alerts
    * Prodes
    * Guira Loss
    * Biomass loss
    * GLAD tree cover loss alerts
    * NASA VIIRs active fires.

    Other services provided by the Api complements the ecosystem that surrounds the platform:

    * Metadata
    * Urthecast
    * Geostore
    * OGR
    * Story
    * User
    * subscription


    ### GeoStore API
    This Microservice will provide a comprehensive way of communicate geometries between the platforms and the microservices.

    We will be able to store a geojson and retrive a geohash for that geojson. The service will also provide information related the geometry stored, like the area in Ha or the bounding box.

    We will also be able to provide a geometry provider at CARTO:
      ```{
      "provider": {
          "filter": "cartodb_id=729417",
          "user": "wri-01",
          "table": "gfw_mining",
          "type": "carto"
      }
    }```


  termsOfService: http://www.globalforestwatch.org/terms
  contact:
    name: API Support
    url: http://www.globalforestwatch.org
    email: support@wri.com
  license:
    name: MIT License (MIT)
    url: https://opensource.org/licenses/MIT
  version: "1.0.0"

host: staging-api.globalforestwatch.org

schemes:
  - https
  - http

consumes:
  - application/vnd.api+json

produces:
  - application/vnd.api+json
###############################################################################
#                                  Security
###############################################################################
#securityDefinitions:
#  UserSecurity:
#    type: apiKey
#   in: header
#    name: Authorization
#    description: For write operations we need to provide a token like  ```Bearer {{TOKEN}}```
#security:
#        - UserSecurity: []
###############################################################################
#                                  Paths
###############################################################################
paths:
  /geostore/{id}:
    get:
      summary: Get geostore by id
      description: |
        Returns geostore based on a single ID, if the geostore not exist return 404
        ### Example URL request
        ```
        {{host}}/geostore/ca38fa80a4ffa9ac6217a7e0bf71e6df
        ```
        ### Example response
        * Saving a geojson
        ```json
        {
              "data": {
                  "type": "geoStore",
                  "id": "ca38fa80a4ffa9ac6217a7e0bf71e6df",
                  "attributes": {
                      "geojson": {
                          "features": [{
                              "type": "Feature",
                              "geometry": {
                                  "type": "Polygon",
                                  "coordinates": [
                                      [
                                          [-5.273512601852417,
                                              42.81137220349083
                                          ],
                                          [-5.273512601852417,
                                              42.811803118457306
                                          ],
                                          [-5.272732079029083,
                                              42.811803118457306
                                          ],
                                          [-5.272732079029083,
                                              42.81137220349083
                                          ],
                                          [-5.273512601852417,
                                              42.81137220349083
                                          ]
                                      ]
                                  ]
                              }
                          }],
                          "crs": {},
                          "type": "FeatureCollection"
                      },
                      "hash": "ca38fa80a4ffa9ac6217a7e0bf71e6df",
                      "provider": {},
                      "areaHa": 0.3057556230214663
                  }
              }
          }
        ```
      operationId: getGeoStoreById
      produces:
        - application/vnd.api+json
      tags:
        - GEOSTORE
      parameters:
        - name: id
          in: path
          description: ID of geostore to fetch ```34cecd4c4809dbbec20ce40be3d1c877```
          required: true
          type: string
      responses:
        200:
          description: Geostore obtained
          schema:
            $ref: '#/definitions/GeoStoreAPI'
        404:
          description: Geostore not found
          schema:
            $ref: '#/definitions/Errors'
        500:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Errors'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Errors'
  /geostore/wdpa/{id}:
    get:
      summary: Get geostore by wdpaid
      description: |
        Returns geostore for a protected are, if the geostore not exist return 404
        ### Example URL request
        ```
        {{host}}/geostore/wdpa/235442
        ```
        ### Example response
        * Saving a geojson
        ```json
        {
              "data": {
                  "type": "geoStore",
                  "id": "ca38fa80a4ffa9ac6217a7e0bf71e6df",
                  "attributes": {
                      "geojson": {
                          "features": [{
                              "type": "Feature",
                              "geometry": {
                                  "type": "Polygon",
                                  "coordinates": [
                                      [
                                          [-5.273512601852417,
                                              42.81137220349083
                                          ],
                                          [-5.273512601852417,
                                              42.811803118457306
                                          ],
                                          [-5.272732079029083,
                                              42.811803118457306
                                          ],
                                          [-5.272732079029083,
                                              42.81137220349083
                                          ],
                                          [-5.273512601852417,
                                              42.81137220349083
                                          ]
                                      ]
                                  ]
                              }
                          }],
                          "crs": {},
                          "type": "FeatureCollection"
                      },
                      "hash": "ca38fa80a4ffa9ac6217a7e0bf71e6df",
                      "provider": {},
                      "areaHa": 0.3057556230214663
                  }
              }
          }
        ```
      operationId: getGeoStoreBywdpaid
      produces:
        - application/vnd.api+json
      tags:
        - GEOSTORE
      parameters:
        - name: id
          in: path
          description: ID of protected area geostore to fetch ```123412```
          required: true
          type: string
      responses:
        200:
          description: Geostore obtained
          schema:
            $ref: '#/definitions/GeoStoreAPI'
        404:
          description: Geostore not found
          schema:
            $ref: '#/definitions/Errors'
        500:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Errors'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Errors'

  /geostore/use/{useType}/{id}:
    get:
      summary: Get geostore by Use id
      description: |
        Returns geostore for a protected are, if the geostore not exist return 404
        ### Example URL request
        ```
        {{host}}/geostore/wdpa/235442
        ```
        ### Example response
        * Saving a geojson
        ```json
        {
              "data": {
                  "type": "geoStore",
                  "id": "ca38fa80a4ffa9ac6217a7e0bf71e6df",
                  "attributes": {
                      "geojson": {
                          "features": [{
                              "type": "Feature",
                              "geometry": {
                                  "type": "Polygon",
                                  "coordinates": [
                                      [
                                          [-5.273512601852417,
                                              42.81137220349083
                                          ],
                                          [-5.273512601852417,
                                              42.811803118457306
                                          ],
                                          [-5.272732079029083,
                                              42.811803118457306
                                          ],
                                          [-5.272732079029083,
                                              42.81137220349083
                                          ],
                                          [-5.273512601852417,
                                              42.81137220349083
                                          ]
                                      ]
                                  ]
                              }
                          }],
                          "crs": {},
                          "type": "FeatureCollection"
                      },
                      "hash": "ca38fa80a4ffa9ac6217a7e0bf71e6df",
                      "provider": {},
                      "areaHa": 0.3057556230214663
                  }
              }
          }
        ```
      operationId: getGeoStoreByUseid
      produces:
        - application/vnd.api+json
      tags:
        - GEOSTORE
      parameters:
        - name: useType
          in: path
          description: Type of concession one of ```mining, oilpalm, fiber, logging```
          required: true
          type: string
        - name: id
          in: path
          description: ID of use area to fetch ```12```
          required: true
          type: string
      responses:
        200:
          description: Geostore obtained
          schema:
            $ref: '#/definitions/GeoStoreAPI'
        404:
          description: Geostore not found
          schema:
            $ref: '#/definitions/Errors'
        500:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Errors'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Errors'

  /geostore:
    post:
      summary: Create geostore
      description: |
        ### 2 posibilities to create geostore
        * Send geojson in a body (geojson attribute)
        * Send provider object with all params. With this mode, the geostore obtain the geojson in the provider. Support providers-> carto

        ### Example response
        * Saving a geojson
        ```json
        {
              "data": {
                  "type": "geoStore",
                  "id": "ca38fa80a4ffa9ac6217a7e0bf71e6df",
                  "attributes": {
                      "geojson": {
                          "features": [{
                              "type": "Feature",
                              "geometry": {
                                  "type": "Polygon",
                                  "coordinates": [
                                      [
                                          [-5.273512601852417,
                                              42.81137220349083
                                          ],
                                          [-5.273512601852417,
                                              42.811803118457306
                                          ],
                                          [-5.272732079029083,
                                              42.811803118457306
                                          ],
                                          [-5.272732079029083,
                                              42.81137220349083
                                          ],
                                          [-5.273512601852417,
                                              42.81137220349083
                                          ]
                                      ]
                                  ]
                              }
                          }],
                          "crs": {},
                          "type": "FeatureCollection"
                      },
                      "hash": "ca38fa80a4ffa9ac6217a7e0bf71e6df",
                      "provider": {},
                      "areaHa": 0.3057556230214663
                  }
              }
          }
        ```
        * Saving with provider
        ```json
        {
          "data": {
              "type": "geoStore",
              "id": "f3ec85259a83c61ac1b7ff8a51d0a461",
              "attributes": {
                  "geojson": {
                      "features": [{
                          "type": "Feature",
                          "geometry": {
                              "type": "MultiPolygon",
                              "coordinates": [
                                  [
                                      [
                                          [-54.3664801812558, -10.9954102656944],
                                          [-54.4224805595865, -10.995410177512],
                                          [-54.4224803350077, -10.8554094476871],
                                          [-54.4134802740762, -10.8554094618029],
                                          [-54.4134802659914, -10.8504094360184],
                                          [-54.3664799476939, -10.8504095083633],
                                          [-54.3664799503888, -10.8514095138597],
                                          [-54.3664801812558, -10.9954102656944]
                                      ]
                                  ]
                              ]
                          }
                      }],
                      "crs": {},
                      "type": "FeatureCollection"
                  },
                  "hash": "f3ec85259a83c61ac1b7ff8a51d0a461",
                  "provider": {
                      "filter": "cartodb_id=729417",
                      "user": "wri-01",
                      "table": "gfw_mining",
                      "type": "carto"
                  },
                  "areaHa": 9764.29591966568
              }
          }
        }
        ```
      consumes:
        - application/json
      produces:
        - application/vnd.api+json
      tags:
        - GEOSTORE
      operationId: createGeoStore
      parameters:
        - name: geoStore
          in: body
          description: |
            ##### Body examples
            * Geojson:
              ```json
              {
                  "geojson": {
                      "type": "FeatureCollection",
                      "features": [{
                          "type": "Feature",
                          "properties": {},
                          "geometry": {
                              "type": "Polygon",
                              "coordinates": [
                                  [
                                      [-5.273512601852417, 42.81137220349083],
                                      [-5.273512601852417, 42.811803118457306],
                                      [-5.272732079029083, 42.811803118457306],
                                      [-5.272732079029083, 42.81137220349083],
                                      [-5.273512601852417, 42.81137220349083]
                                  ]
                              ]
                          }
                      }]
                  }
              }
              ```
            * Provider:
              ```json
                {
                  "provider": {
                      "filter": "cartodb_id=729417",
                      "user": "wri-01",
                      "table": "gfw_mining",
                      "type": "carto"
                  }
                }
              ```
          required: true
          schema:
            $ref: '#/definitions/GeoStoreBodyRequest'
      responses:
        200:
          description: The object created
          schema:
            $ref: '#/definitions/GeoStoreAPI'
        400:
          description: |
            Validate error.
            * ProviderNotFound-> If the provider is distinct to 'carto', the api return this error
            * GeoJSONNotFound -> If the microservice try to obtain the geojson in the provider and the provider return empty response, the microservice return this error.
          schema:
            $ref: '#/definitions/Errors'
        500:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Errors'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Errors'

definitions:
  Errors:
    type: object
    description: Errors
    properties:
      errors:
        type: array
        items:
          $ref: '#/definitions/Error'

  Error:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: A unique identifier for this particular occurrence of the problem.
      links:
        type: object
        description: A links object
        properties:
          about:
            type: string
            description: A link that leads to further details about this particular occurrence of the problem.
      status:
        type: string
        description: The HTTP status code applicable to this problem, expressed as a string value
      code:
        type: string
        description: An application-specific error code, expressed as a string value
      title:
        type: string
        description: A short, human-readable summary of the problem that SHOULD NOT change from occurrence to occurrence of the problem, except for purposes of localization.
      detail:
        type: string
        description: A human-readable explanation specific to this occurrence of the problem. Like title, this field's value can be localized
      source:
        type: object
        description: An object containing references to the source of the error, optionally including any of the following members
        properties:
          pointer:
            type: string
            description: A JSON Pointer [RFC6901] to the associated entity in the request document
          parameter:
            type: string
            description: A string indicating which URI query parameter caused the error.
      meta:
        type: object
        description: A meta object containing non-standard meta-information about the error.

  GeoStoreBodyRequest:
    type: object
    properties:
      provider:
        type: object
        description: Data where obtain geojson
        properties:
          type:
            type: string
            description: type of provider (posibilities-> carto)
          table:
            type: string
            description: table name
          user:
            type: string
            description: owner of table
          filter:
            type: string
            description: filter to obtain the geostore
        required:
          - type
          - table
          - user
          - filter
        example:
          provider:
            filter: cartodb_id=729417
            user: wri-01
            table: gfw_mining
            type: carto
      geojson:
        properties:
          schema:
            $ref: '#/definitions/Features'

  GeoStoreAPI:
    type: object
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/GeoStore'

  GeoStore:
    type: object
    properties:
      type:
        type: string
        description: Resource type
      id:
        type: string
        description: Resource id
      attributes:
        properties:
          schema:
            $ref: '#/definitions/GeoStoreBodyResponse'
  GeoStoreBodyResponse:
    type: object
    properties:
      hash:
        type: string
        description: Resource type
      areaHa:
        type: string
        description: Area in
      provider:
        type: object
        description: Data where obtain geojson
        properties:
          type:
            type: string
            description: type of provider (posibilities-> carto)
          table:
            type: string
            description: table name
          user:
            type: string
            description: owner of table
          filter:
            type: string
            description: filter to obtain the geostore
      geojson:
        properties:
          schema:
            $ref: '#/definitions/Features'
  Features:
    type: object
    properties:
      type:
        type: string
        description: Type of feature
      features:
        type: array
        description: Features
        items:
          $ref: '#/definitions/Feature'
  Feature:
    type: object
    properties:
      type:
        type: string
        description: Type of feature
      properties:
        type: object
        description: Feature
      geometry:
        type: object
        description: Geometries
        properties:
          schema:
            $ref: '#/definitions/Geometry'

  Geometry:
    type: object
    properties:
      type:
        type: string
        description: Geometry type
      coordinates:
        type: array
        description: Coordinates array
        items:
          type: object